requirements.txt
backports.tarfile==1.2.0
bertviz==1.4.0
brotli==1.0.9
cached-property==1.5.2
circuitsvis==1.43.3
en-core-web-sm==3.8.0
evaluate==0.4.6
exceptiongroup==1.3.0
fqdn==1.5.1
gcsfs==2025.3.0
greenlet==3.2.2
h2==4.2.0
imgkit==1.2.3
importlib-resources==6.5.2
ipywidgets==8.1.7
isoduration==20.11.0
jaraco.collections==5.1.0
langchain-community==0.3.24
langchain-nvidia-ai-endpoints==0.3.9
llama-index==0.12.35
matplotlib==3.10.3
neel-plotly==0.0.0
notebook==7.4.2
peft==0.17.1
pickleshare==0.7.5
pkgutil-resolve-name==1.3.10
pyobjc-framework-cocoa==10.1
pysocks==1.7.1
rouge-score==0.1.2
scikit-learn==1.7.1
spacy==3.8.7
tinycss2==1.4.0
tomli==2.2.1
torchrl==0.7.2
torchtyping==0.1.5
transformer-lens==0.0.0
trl==0.17.0
typing-utils==0.1.0
uri-template==1.3.0
wadler-lindig==0.1.7
webcolors==24.11.1
wikipedia==1.4.0
